RapidPrototyping/requirements.txt
# Core dependencies
openai>=1.0.0
clarifai>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
markdown>=3.4.0
python-frontmatter>=1.0.0

# Data handling
pyyaml>=6.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0

# CLI and UI
rich>=13.0.0
typer>=0.9.0
questionary>=2.0.0

# Async support
aiohttp>=3.8.0
aiofiles>=23.0.0

# Utilities
pathlib2>=2.3.0
watchdog>=3.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
